chore:#264 회원 삭제 기능 역할 분리
