+++ src/main/java/or/sopt/houme/domain/user/service/UserService.java
@@ -20,5 +20,4 @@ public interface UserService {
     // 사용자 이미지 생성 여부 저장
     void updateHasGeneratedImage(User user);
 
-    void delete(Long userId);
 }
